+++ Source/ACE/Entity/SpellLikeEffect.cs
@@ -1,9 +1,4 @@
 ﻿using ACE.Network.Enum;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using ACE.Entity.Enum;
 using log4net;
 using ACE.Network.GameMessages.Messages;
@@ -14,9 +9,9 @@ namespace ACE.Entity
     public class SpellLikeEffect : WorldObject
     {
         private static readonly ILog log = LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
-        
+
         public PlayScript PlayerScript = Network.Enum.PlayScript.Invalid;
-        
+
         public SpellLikeEffect(ObjectType type,
             ObjectGuid guid,
             string name,
@@ -29,23 +24,23 @@ namespace ACE.Entity
             uint soundTableId,
             uint physicsTableId) : base(type, guid)
         {
-            this.Name = name;
-            this.DescriptionFlags = descriptionFlag;
-            this.WeenieFlags = weenieFlag;
-            this.Location = position;
-            this.WeenieClassid = weenieClassId;
+            Name = name;
+            DescriptionFlags = descriptionFlag;
+            WeenieFlags = weenieFlag;
+            Location = position;
+            WeenieClassid = weenieClassId;
 
-            this.Spell = spellId;
-            this.PhysicsData.CSetup = modelId;
-            this.PhysicsData.Stable = soundTableId;
-            this.PhysicsData.Petable = physicsTableId;
+            Spell = spellId;
+            SetupTableId = modelId;
+            SoundTableId = soundTableId;
+            PhisicsTableId = physicsTableId;
         }
 
         public override void PlayScript(Session session)
         {
             if (PlayerScript != Network.Enum.PlayScript.Invalid)
             {
-                var scriptEvent = new GameMessageScript(this.Guid, PlayerScript, 1f);
+                var scriptEvent = new GameMessageScript(Guid, PlayerScript, 1f);
                 session.Network.EnqueueSend(scriptEvent);
             }
         }