+++ Source/ACE/Entity/WorldObject.cs
@@ -12,6 +12,8 @@ using System.IO;
 using ACE.Managers;
 using log4net;
 using System;
+using System.Diagnostics;
+using System.Linq;
 
 namespace ACE.Entity
 {
@@ -31,8 +33,8 @@ namespace ACE.Entity
 
         public ObjectType Type
         {
-            get { return (ObjectType)AceObject.ItemType; }
-            protected set { AceObject.ItemType = (uint)value; }
+            get { return (ObjectType)AceObject.Type; }
+            protected set { AceObject.Type = (uint)value; }
         }
 
         /// <summary>
@@ -64,7 +66,8 @@ namespace ACE.Entity
         /// <summary>
         /// Should only be adjusted by LandblockManager -- default is null
         /// </summary>
-        public Landblock CurrentLandblock {
+        public Landblock CurrentLandblock
+        {
             get
             {
                 return CurrentParent as Landblock;
@@ -89,15 +92,13 @@ namespace ACE.Entity
             set
             {
                 /*
-                log.Debug($"{Name} moved to {PhysicsData.Position}");
+                log.Debug($"{Name} moved to {Position}");
 
-                PhysicsData.Position = value;
+                Position = value;
                 */
                 if (AceObject.Location != null)
                     LastUpdatedTicks = WorldManager.PortalYearTicks;
                 AceObject.Location = value;
-                // FIXME(ddevec): When PhysicsData is factored out this can be deleted
-                PhysicsData.Position = value;
             }
         }
 
@@ -112,7 +113,11 @@ namespace ACE.Entity
         /// </summary>
         public double LastAnimatedTicks { get; set; }
 
-        public ObjectDescriptionFlag DescriptionFlags { get; protected set; }
+        public ObjectDescriptionFlag DescriptionFlags
+        {
+            get { return (ObjectDescriptionFlag)AceObject.AceObjectDescriptionFlags; }
+            protected internal set { AceObject.AceObjectDescriptionFlags = (uint)value; }
+        }
 
         public WeenieHeaderFlag WeenieFlags
         {
@@ -130,15 +135,11 @@ namespace ACE.Entity
 
         public virtual float ListeningRadius { get; protected set; } = 5f;
 
-        public ModelData ModelData { get; }
-
-        public PhysicsData PhysicsData { get; }
-
         // Logical Game Data
         public uint GameDataType
         {
-            get { return AceObject.ItemType; }
-            set { AceObject.ItemType = value; }
+            get { return AceObject.Type; }
+            set { AceObject.Type = value; }
         }
 
         public ContainerType ContainerType
@@ -309,12 +310,7 @@ namespace ACE.Entity
         public Spell? Spell
         {
             get { return (Spell?)AceObject.SpellId; }
-            set { this.AceObject.SpellId = (ushort?)value; }
-        }
-        public uint? ClothingBase
-        {
-            get { return AceObject.ClothingBase; }
-            set { AceObject.ClothingBase = value; }
+            set { AceObject.SpellId = (ushort?)value; }
         }
 
         /// <summary>
@@ -358,8 +354,7 @@ namespace ACE.Entity
 
         public uint? PetOwner { get; set; }
 
-        // WeenieHeaderFlag2
-        public uint? Cooldown
+       public uint? CooldownId
         {
             get { return AceObject.CooldownId; }
             set { AceObject.CooldownId = value; }
@@ -372,49 +367,191 @@ namespace ACE.Entity
         }
 
         // PhysicsData Logical
+        /// <summary>
+        /// setup_id in aclogviewer - used to get the correct model out of the dat file
+        /// </summary>
+        public uint? SetupTableId
+        {
+            get { return AceObject.SetupTableId; }
+            set { AceObject.SetupTableId = value; }
+        }
 
-        public uint CSetup;
+        public PhysicsDescriptionFlag PhysicsDescriptionFlag { get; set; }
 
-        // apply default for back compat with player object
-        // public PhysicsDescriptionFlag PhysicsDescriptionFlag;
-        // public PhysicsState PhysicsState = 0;
+        public PhysicsState PhysicsState
+        {
+            get { return (PhysicsState)AceObject.PhysicsState; }
+            set { AceObject.PhysicsState = (uint)value; }
+        }
+       
+        /// <summary>
+        /// mtable_id in aclogviewer This is the sound table for the object.   Looked up from dat file.
+        /// </summary>
+        public uint? MotionTableId
+        {
+            get { return AceObject.MotionTableId; }
+            set { AceObject.MotionTableId = value; }
+        }
+        /// <summary>
+        /// stable_id in aclogviewer This is the sound table for the object.   Looked up from dat file.
+        /// </summary>
+        public uint? SoundTableId
+        {
+            get { return AceObject.SoundTableId; }
+            set { AceObject.SoundTableId = value; }
+        }
+        /// <summary>
+        /// phstable_id in aclogviewer This is the physics table for the object.   Looked up from dat file.
+        /// </summary>
+        public uint? PhisicsTableId
+        {
+            get { return AceObject.PhysicsTableId; }
+            set { AceObject.PhysicsTableId = value; }
+        }
 
-        // public Position Position;
+        public uint ItemsEquipedCount { get; set; }
+        /// <summary>
+        /// This is used for equiped items that are selectable.   Weapons or shields only.   Max 2
+        /// </summary>
+        public uint? Parent
+        {
+            get { return AceObject.Parent; }
+            set { AceObject.Parent = value; }
+        }
 
-        // these are all related
+        public uint? ParentLocation
+        {
+            get { return AceObject.ParentLocation; }
+            set { AceObject.ParentLocation = value; }
+        }
+
         public EquipMask? EquipperPhysicsDescriptionFlag;
-        private readonly List<EquippedItem> children = new List<EquippedItem>();
 
-        public float? ObjScale;
-        public float? Friction;
-        public float? Elasticity;
-        public uint? AnimationFrame;
-        public AceVector3 Acceleration;
-        public float? Translucency;
+        public List<EquippedItem> Children { get; } = new List<EquippedItem>();
+
+        public float? ObjScale
+        {
+            get { return AceObject.DefaultScale; }
+            set { AceObject.DefaultScale = value; }
+        }
+
+        public float? Friction
+        {
+            get { return AceObject.Friction; }
+            set { AceObject.Friction = value; }
+        }
+
+        public float? Elasticity
+        {
+            get { return AceObject.Elasticity; }
+            set { AceObject.Elasticity = value; }
+        }
+
+        public uint? AnimationFrame
+        {
+            get { return AceObject.AnimationFrameId; }
+            set { AceObject.AnimationFrameId = value; }
+        }
+
+        public AceVector3 Acceleration { get; set; }
+
+        public float? Translucency
+        {
+            get { return AceObject.Translucency; }
+            set { AceObject.Translucency = value; }
+        }
+
         public AceVector3 Velocity = null;
+
         public AceVector3 Omega = null;
 
         private MotionState currentMotionState;
+
         public MotionState CurrentMotionState
         {
             get { return currentMotionState; }
             set { currentMotionState = value; }
         }
 
-        public uint? DefaultScript;
-        public float? DefaultScriptIntensity;
+        public uint? DefaultScript
+        {
+            get { return AceObject.DefaultScript; }
+            set { AceObject.DefaultScript = value; }
+        }
+
+        public float? DefaultScriptIntensity
+        {
+            get { return AceObject.PhysicsScriptIntensity; }
+            set { AceObject.PhysicsScriptIntensity = value; }
+        }
+
+        // START of Logical Model Data
+
+        public uint? PaletteGuid
+        {
+            get { return AceObject.PaletteId; }
+            set { AceObject.PaletteId = value; }
+        }
+
+        public uint? ClothingBase
+        {
+            get { return AceObject.ClothingBase; }
+            set { AceObject.ClothingBase = value; }
+        }
+
+        private readonly List<ModelPalette> modelPalettes = new List<ModelPalette>();
+
+        private readonly List<ModelTexture> modelTextures = new List<ModelTexture>();
+
+        private readonly List<Model> models = new List<Model>();
+
+        public List<ModelPalette> GetPalettes
+        {
+            get { return modelPalettes.ToList(); }
+        }
+
+        public List<ModelTexture> GetTextures
+        {
+            get { return modelTextures.ToList(); }
+        }
+
+        public List<Model> GetModels
+        {
+            get { return models.ToList(); }
+        }
+
+        public void AddPalette(uint paletteId, ushort offset, ushort length)
+        {
+            ModelPalette newpalette = new ModelPalette(paletteId, offset, length);
+            modelPalettes.Add(newpalette);
+        }
+
+        public void AddTexture(byte index, uint oldtexture, uint newtexture)
+        {
+            ModelTexture nextTexture = new ModelTexture(index, oldtexture, newtexture);
+            modelTextures.Add(nextTexture);
+        }
+
+        public void AddModel(byte index, uint modelresourceid)
+        {
+            var newmodel = new Model(index, modelresourceid);
+            models.Add(newmodel);
+        }
+
+        public void Clear()
+        {
+            modelPalettes.Clear();
+            modelTextures.Clear();
+            models.Clear();
+        }
 
         public SequenceManager Sequences { get; }
 
         protected WorldObject(ObjectType type, ObjectGuid guid)
         {
-            // Kludge until I get everything refactored correctly Og II
-            if (AceObject == null)
-                AceObject = new AceObject(guid.Full);
+            AceObject = new AceObject();
             Type = type;
             Guid = guid;
-
-            ModelData = new ModelData();
 
             Sequences = new SequenceManager();
             Sequences.AddOrSetSequence(SequenceType.ObjectPosition, new UShortSequence());
@@ -427,14 +564,45 @@ namespace ACE.Entity
             Sequences.AddOrSetSequence(SequenceType.ObjectVisualDesc, new UShortSequence());
             Sequences.AddOrSetSequence(SequenceType.ObjectInstance, new UShortSequence());
             Sequences.AddOrSetSequence(SequenceType.Motion, new UShortSequence(1));
-
-            PhysicsData = new PhysicsData(Sequences);
         }
 
         protected WorldObject(AceObject aceObject)
-                : this((ObjectType)aceObject.ItemType, new ObjectGuid(aceObject.AceObjectId))
+                : this((ObjectType)aceObject.Type, new ObjectGuid(aceObject.AceObjectId))
         {
-            this.AceObject = aceObject;
+            AceObject = aceObject;
+            if (aceObject.CurrentMotionState == "0" || aceObject.CurrentMotionState == null)
+                CurrentMotionState = null;
+            else
+                CurrentMotionState = new UniversalMotion(Convert.FromBase64String(aceObject.CurrentMotionState));
+
+            SetPhysicsDescriptionFlag(this);
+            WeenieFlags = SetWeenieHeaderFlag();
+            WeenieFlags2 = SetWeenieHeaderFlag2();
+
+            aceObject.AnimationOverrides.ForEach(ao => AddModel(ao.Index, ao.AnimationId));
+            aceObject.TextureOverrides.ForEach(to => AddTexture(to.Index, to.OldId, to.NewId));
+            aceObject.PaletteOverrides.ForEach(po => AddPalette(po.SubPaletteId, po.Offset, po.Length));
+            PaletteGuid = aceObject.PaletteId;
+        }
+
+        protected WorldObject(ObjectGuid guid, AceObject aceObject)
+            : this((ObjectType)aceObject.Type, guid)
+        {
+            Guid = guid;
+            AceObject = aceObject;
+            if (aceObject.CurrentMotionState == "0" || aceObject.CurrentMotionState == null)
+                CurrentMotionState = null;
+            else
+                CurrentMotionState = new UniversalMotion(Convert.FromBase64String(aceObject.CurrentMotionState));
+
+            SetPhysicsDescriptionFlag(this);
+            WeenieFlags = SetWeenieHeaderFlag();
+            WeenieFlags2 = SetWeenieHeaderFlag2();
+
+            aceObject.AnimationOverrides.ForEach(ao => AddModel(ao.Index, ao.AnimationId));
+            aceObject.TextureOverrides.ForEach(to => AddTexture(to.Index, to.OldId, to.NewId));
+            aceObject.PaletteOverrides.ForEach(po => AddPalette(po.SubPaletteId, po.Offset, po.Length));
+            PaletteGuid = aceObject.PaletteId;
         }
 
         public void SetCombatMode(CombatMode newCombatMode)
@@ -457,13 +625,13 @@ namespace ACE.Entity
 
         internal void SetInventoryForWorld(WorldObject inventoryItem)
         {
-            inventoryItem.Location = PhysicsData.Position.InFrontOf(1.1f);
+            inventoryItem.Location = Location.InFrontOf(1.1f);
             inventoryItem.PositionFlag = UpdatePositionFlag.Contact
                                          | UpdatePositionFlag.Placement
                                          | UpdatePositionFlag.ZeroQy
                                          | UpdatePositionFlag.ZeroQx;
 
-            inventoryItem.PhysicsData.PhysicsDescriptionFlag = inventoryItem.PhysicsData.SetPhysicsDescriptionFlag(inventoryItem);
+            inventoryItem.PhysicsDescriptionFlag = inventoryItem.SetPhysicsDescriptionFlag(inventoryItem);
             inventoryItem.ContainerId = null;
             inventoryItem.Wielder = null;
             inventoryItem.WeenieFlags = inventoryItem.SetWeenieHeaderFlag();
@@ -473,9 +641,8 @@ namespace ACE.Entity
         {
             if (inventoryItem.Location != null)
                 LandblockManager.RemoveObject(inventoryItem);
-            inventoryItem.PhysicsData.PhysicsDescriptionFlag &= ~PhysicsDescriptionFlag.Position;
-            inventoryItem.PositionFlag = UpdatePositionFlag.None;
-            inventoryItem.PhysicsData.Position = null;
+            inventoryItem.PhysicsDescriptionFlag &= ~PhysicsDescriptionFlag.Position;
+            inventoryItem.PositionFlag = UpdatePositionFlag.None;            
             inventoryItem.Location = null;
             inventoryItem.WeenieFlags = inventoryItem.SetWeenieHeaderFlag();
         }
@@ -483,7 +650,7 @@ namespace ACE.Entity
         public void SetMotionState(MotionState motionState)
         {
             var p = (Player)this;
-            PhysicsData.CurrentMotionState = motionState;
+            CurrentMotionState = motionState;
             var updateMotion = new GameMessageUpdateMotion(p.Guid,
                                                            p.Sequences.GetCurrentSequence(SequenceType.ObjectInstance),
                                                            p.Sequences, motionState);
@@ -612,7 +779,7 @@ namespace ACE.Entity
             if ((IconUnderlay != null) && (IconUnderlay != 0))
                 weenieHeaderFlag2 |= WeenieHeaderFlag2.IconUnderlay;
 
-            if ((Cooldown != null) && (Cooldown != 0))
+            if ((CooldownId != null) && (CooldownId != 0))
                 weenieHeaderFlag2 |= WeenieHeaderFlag2.Cooldown;
 
             if ((CooldownDuration != null) && Math.Abs((float)CooldownDuration) >= 0.001)
@@ -628,8 +795,8 @@ namespace ACE.Entity
         {
             writer.WriteGuid(Guid);
 
-            ModelData.Serialize(writer);
-            PhysicsData.Serialize(this, writer);
+            SerializeModelData(writer);
+            Serialize(this, writer);
             WeenieFlags = SetWeenieHeaderFlag();
             WeenieFlags2 = SetWeenieHeaderFlag2();
             if (WeenieFlags2 != WeenieHeaderFlag2.None)
@@ -747,7 +914,7 @@ namespace ACE.Entity
                 writer.Write((uint)(Material ?? 0u));
 
             if ((WeenieFlags2 & WeenieHeaderFlag2.Cooldown) != 0)
-                writer.Write(Cooldown ?? 0u);
+                writer.Write(CooldownId ?? 0u);
 
             if ((WeenieFlags2 & WeenieHeaderFlag2.CooldownDuration) != 0)
                 writer.Write((double?)CooldownDuration ?? 0u);
@@ -765,9 +932,41 @@ namespace ACE.Entity
         public virtual void SerializeUpdateModelData(BinaryWriter writer)
         {
             writer.WriteGuid(Guid);
-            ModelData.Serialize(writer);
+            SerializeModelData(writer);
             writer.Write(Sequences.GetCurrentSequence(SequenceType.ObjectInstance));
             writer.Write(Sequences.GetNextSequence(SequenceType.ObjectPosition));
+        }
+
+        public void SerializeModelData(BinaryWriter writer)
+        {
+            writer.Write((byte)0x11);
+            writer.Write((byte)modelPalettes.Count);
+            writer.Write((byte)modelTextures.Count);
+            writer.Write((byte)models.Count);
+
+            if ((modelPalettes.Count > 0) && (PaletteGuid != null))
+                writer.WritePackedDwordOfKnownType((uint)PaletteGuid, 0x4000000);
+            foreach (var palette in modelPalettes)
+            {
+                writer.WritePackedDwordOfKnownType(palette.PaletteId, 0x4000000);
+                writer.Write((byte)palette.Offset);
+                writer.Write((byte)palette.Length);
+            }
+
+            foreach (var texture in modelTextures)
+            {
+                writer.Write((byte)texture.Index);
+                writer.WritePackedDwordOfKnownType(texture.OldTexture, 0x5000000);
+                writer.WritePackedDwordOfKnownType(texture.NewTexture, 0x5000000);
+            }
+
+            foreach (var model in models)
+            {
+                writer.Write((byte)model.Index);
+                writer.WritePackedDwordOfKnownType(model.ModelID, 0x1000000);
+            }
+
+            writer.Align();
         }
 
         public void WriteUpdatePositionPayload(BinaryWriter writer)
@@ -867,6 +1066,181 @@ namespace ACE.Entity
         public void RunActions()
         {
             actionQueue.RunActions();
+        }
+
+        public PhysicsDescriptionFlag SetPhysicsDescriptionFlag(WorldObject wo)
+        {
+            var physicsDescriptionFlag = PhysicsDescriptionFlag.None;
+
+            if (CurrentMotionState != null)
+                physicsDescriptionFlag |= PhysicsDescriptionFlag.Movement;
+
+            if ((AnimationFrame != null) && (AnimationFrame != 0))
+                physicsDescriptionFlag |= PhysicsDescriptionFlag.AnimationFrame;
+
+            if (Location != null)
+                physicsDescriptionFlag |= PhysicsDescriptionFlag.Position;
+
+            // NOTE: While we fill with 0 the flag still has to reflect that we are not really making this entry for the client.
+            if (MotionTableId != 0)
+                physicsDescriptionFlag |= PhysicsDescriptionFlag.MTable;
+
+            if (SoundTableId != 0)
+                physicsDescriptionFlag |= PhysicsDescriptionFlag.Stable;
+
+            if (PhisicsTableId != 0)
+                physicsDescriptionFlag |= PhysicsDescriptionFlag.Petable;
+
+            if (SetupTableId != 0)
+                physicsDescriptionFlag |= PhysicsDescriptionFlag.CSetup;
+
+            if (ItemsEquipedCount != 0)
+                physicsDescriptionFlag |= PhysicsDescriptionFlag.Children;
+
+            if (Parent != null)
+                physicsDescriptionFlag |= PhysicsDescriptionFlag.Parent;
+
+            if ((ObjScale != null) && (Math.Abs((float)ObjScale) >= 0.001))
+                physicsDescriptionFlag |= PhysicsDescriptionFlag.ObjScale;
+
+            if (Friction != null)
+                physicsDescriptionFlag |= PhysicsDescriptionFlag.Friction;
+
+            if (Elasticity != null)
+                physicsDescriptionFlag |= PhysicsDescriptionFlag.Elasticity;
+
+            if ((Translucency != null) && (Math.Abs((float)Translucency) >= 0.001))
+                physicsDescriptionFlag |= PhysicsDescriptionFlag.Translucency;
+
+            if (Velocity != null)
+                physicsDescriptionFlag |= PhysicsDescriptionFlag.Velocity;
+
+            if (Acceleration != null)
+                physicsDescriptionFlag |= PhysicsDescriptionFlag.Acceleration;
+
+            if (Omega != null)
+                physicsDescriptionFlag |= PhysicsDescriptionFlag.Omega;
+
+            if (DefaultScript != null)
+                physicsDescriptionFlag |= PhysicsDescriptionFlag.DefaultScript;
+
+            if (DefaultScriptIntensity != null)
+                physicsDescriptionFlag |= PhysicsDescriptionFlag.DefaultScriptIntensity;
+
+            return physicsDescriptionFlag;
+        }
+        // todo: return bytes of data for network write ? ?
+        public void Serialize(WorldObject wo, BinaryWriter writer)
+        {
+            PhysicsDescriptionFlag = SetPhysicsDescriptionFlag(wo);
+
+            writer.Write((uint)PhysicsDescriptionFlag);
+
+            writer.Write((uint)PhysicsState);
+
+            if ((PhysicsDescriptionFlag & PhysicsDescriptionFlag.Movement) != 0)
+            {
+                if (CurrentMotionState != null)
+                {
+                    var movementData = CurrentMotionState.GetPayload(wo.Guid, wo.Sequences);
+                    writer.Write(movementData.Length); // number of bytes in movement object
+                    writer.Write(movementData);
+                    uint autonomous = CurrentMotionState.IsAutonomous ? (ushort)1 : (ushort)0;
+                    writer.Write(autonomous);
+                }
+                else // create a new current motion state and send it.
+                {
+                    CurrentMotionState = new UniversalMotion(MotionStance.Standing);
+                    var movementData = CurrentMotionState.GetPayload(wo.Guid, wo.Sequences);
+                    writer.Write(movementData.Length);
+                    writer.Write(movementData);
+                    uint autonomous = CurrentMotionState.IsAutonomous ? (ushort)1 : (ushort)0;
+                    writer.Write(autonomous);
+                }
+            }
+
+            if ((PhysicsDescriptionFlag & PhysicsDescriptionFlag.AnimationFrame) != 0)
+                writer.Write((AnimationFrame ?? 0u));
+
+            if ((PhysicsDescriptionFlag & PhysicsDescriptionFlag.Position) != 0)
+                Location.Serialize(writer);
+
+            if ((PhysicsDescriptionFlag & PhysicsDescriptionFlag.MTable) != 0)
+                writer.Write(MotionTableId ?? 0u);
+
+            // stable_id =  BYTE1(v12) & 8 )  =  8
+            if ((PhysicsDescriptionFlag & PhysicsDescriptionFlag.Stable) != 0)
+                writer.Write(SoundTableId ?? 0u);
+
+            // setup id
+            if ((PhysicsDescriptionFlag & PhysicsDescriptionFlag.Petable) != 0)
+                writer.Write(PhisicsTableId ?? 0u);
+
+            if ((PhysicsDescriptionFlag & PhysicsDescriptionFlag.CSetup) != 0)
+                writer.Write(SetupTableId ?? 0u);
+
+            if ((PhysicsDescriptionFlag & PhysicsDescriptionFlag.Parent) != 0)
+            {
+                writer.Write((uint)Parent);
+                writer.Write((uint)ParentLocation);
+            }
+
+            if ((PhysicsDescriptionFlag & PhysicsDescriptionFlag.Children) != 0)
+            {
+                writer.Write(ItemsEquipedCount);
+                foreach (var child in Children)
+                {
+                    writer.Write(child.Guid);
+                    writer.Write(1u); // This is going to be child.ParentLocation when we get to it
+                }
+            }
+
+            if ((PhysicsDescriptionFlag & PhysicsDescriptionFlag.ObjScale) != 0)
+                writer.Write(ObjScale ?? 0u);
+
+            if ((PhysicsDescriptionFlag & PhysicsDescriptionFlag.Friction) != 0)
+                writer.Write(Friction ?? 0u);
+
+            if ((PhysicsDescriptionFlag & PhysicsDescriptionFlag.Elasticity) != 0)
+                writer.Write(Elasticity ?? 0u);
+
+            if ((PhysicsDescriptionFlag & PhysicsDescriptionFlag.Translucency) != 0)
+                writer.Write(Translucency ?? 0u);
+
+            if ((PhysicsDescriptionFlag & PhysicsDescriptionFlag.Velocity) != 0)
+            {
+                Debug.Assert(Velocity != null, "Velocity != null");
+                // We do a null check above and unset the flag so this has to be good.
+                Velocity.Serialize(writer);
+            }
+
+            if ((PhysicsDescriptionFlag & PhysicsDescriptionFlag.Acceleration) != 0)
+            {
+                Acceleration.Serialize(writer);
+            }
+
+            if ((PhysicsDescriptionFlag & PhysicsDescriptionFlag.Omega) != 0)
+            {
+                Omega.Serialize(writer);
+            }
+
+            if ((PhysicsDescriptionFlag & PhysicsDescriptionFlag.DefaultScript) != 0)
+                writer.Write(DefaultScript ?? 0u);
+
+            if ((PhysicsDescriptionFlag & PhysicsDescriptionFlag.DefaultScriptIntensity) != 0)
+                writer.Write(DefaultScriptIntensity ?? 0u);
+
+            writer.Write(Sequences.GetCurrentSequence(SequenceType.ObjectPosition));
+            writer.Write(Sequences.GetCurrentSequence(SequenceType.ObjectMovement));
+            writer.Write(Sequences.GetCurrentSequence(SequenceType.ObjectState));
+            writer.Write(Sequences.GetCurrentSequence(SequenceType.ObjectVector));
+            writer.Write(Sequences.GetCurrentSequence(SequenceType.ObjectTeleport));
+            writer.Write(Sequences.GetCurrentSequence(SequenceType.ObjectServerControl));
+            writer.Write(Sequences.GetCurrentSequence(SequenceType.ObjectForcePosition));
+            writer.Write(Sequences.GetCurrentSequence(SequenceType.ObjectVisualDesc));
+            writer.Write(Sequences.GetCurrentSequence(SequenceType.ObjectInstance));
+
+            writer.Align();
         }
     }
 }