+++ Source/ACE/Network/Handlers/CharacterHandler.cs
@@ -137,7 +137,7 @@ namespace ACE.Network.Handlers
             character.MotionTableId = sex.MotionTable;
             character.SoundTableId = sex.SoundTable;
             character.PhysicsTableId = sex.PhysicsTable;
-            character.ModelTableId = sex.SetupID;
+            character.SetupTableId = sex.SetupID;
             character.PaletteId = sex.BasePalette;
             character.CombatTableId = sex.CombatTable;
 
@@ -153,7 +153,7 @@ namespace ACE.Network.Handlers
 
             // Certain races (Undead, Tumeroks, Others?) have multiple body styles available. This is controlled via the "hair style".
             if (hairstyle.AlternateSetup > 0)
-                character.ModelTableId = hairstyle.AlternateSetup;
+                character.SetupTableId = hairstyle.AlternateSetup;
 
             character.EyesTexture = sex.GetEyeTexture(appearance.Eyes, isBald);
             character.DefaultEyesTexture = sex.GetDefaultEyeTexture(appearance.Eyes, isBald);
@@ -169,18 +169,18 @@ namespace ACE.Network.Handlers
             PaletteSet skinPalSet = PaletteSet.ReadFromDat(sex.SkinPalSet);
             ushort skinPal = (ushort)skinPalSet.GetPaletteID(appearance.SkinHue);
             character.SkinPalette = skinPal;
-            // ModelData.AddPalette(skinPal, 0x0, 0x18); // for reference on how to apply
+            // AddPalette(skinPal, 0x0, 0x18); // for reference on how to apply
 
             // Hair is stored as PaletteSet (list of Palettes), so we need to read in the set to get the specific palette
             PaletteSet hairPalSet = PaletteSet.ReadFromDat(sex.HairColorList[Convert.ToInt32(appearance.HairColor)]);
             ushort hairPal = (ushort)hairPalSet.GetPaletteID(appearance.HairHue);
 
             character.HairPalette = hairPal;
-            // ModelData.AddPalette(hairPal, 0x18, 0x8); // for reference on how to apply
+            // AddPalette(hairPal, 0x18, 0x8); // for reference on how to apply
 
             // Eye Color
             character.EyesPalette = sex.EyeColorList[Convert.ToInt32(appearance.EyeColor)];
-            // ModelData.AddPalette(PropertyDataId.EyesPalette, 0x20, 0x8); // for reference on how to apply
+            // AddPalette(PropertyDataId.EyesPalette, 0x20, 0x8); // for reference on how to apply
 
             if (appearance.HeadgearStyle < 0xFFFFFFFF) // No headgear is max UINT
             {